backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
insightface
onnxruntime
faiss-cpu
numpy
Pillow
opencv-python
